# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS API",
        "description": "Laravel CMS REST API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "List pages",
                "description": "Returns a paginated list of pages. Supports search, status filter, menu filter, and pagination.",
                "operationId": "374d995436029ca3d249e45ee40dcd70",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term matched against page title/body",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "about"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by page status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "published"
                            ]
                        }
                    },
                    {
                        "name": "menu_id",
                        "in": "query",
                        "description": "Filter by menu ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of pages"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Create a page",
                "operationId": "a6584a2336ff5430afc72ba1058c4b92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "menu_id",
                                    "title",
                                    "body",
                                    "status"
                                ],
                                "properties": {
                                    "menu_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "About Us"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "This is our about page."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft",
                                        "enum": [
                                            "draft",
                                            "published"
                                        ]
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pages/{page}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Show a single page",
                "operationId": "b70e77e087b972c58f9b90b53ef12dc3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update a page",
                "operationId": "7fb9cc8c654156494901b7d2ebdaf064",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "menu_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Delete a page",
                "operationId": "c6d46df85e358b0fb83bd91d29282de0",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pages/{page}/restore": {
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Restore a soft deleted page",
                "operationId": "99613a0d25dcf374a0c15a647c77a2dc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page restored successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/pages/trash": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "List soft deleted pages",
                "operationId": "c652e914b92205f0301682e517d2f6c7",
                "responses": {
                    "200": {
                        "description": "List of soft deleted pages"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pages/{id}/force": {
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Permanently delete a soft deleted page",
                "operationId": "b33713b352ad1d75a1d0203d852981e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page permanently deleted"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum",
                "description": "Enter your Bearer token"
            }
        }
    },
...
}